pyproject.toml
[tool.poetry]
name = "advanced library system"
version = "0.1.0"
description = ""
authors = ["Mohamed Zakaria <mohamedzakaria20001@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
djangorestframework = "^3.16.0"
djangorestframework-simplejwt = "^5.5.0"
django-filter = "^25.1"
celery = "^5.5.2"
django-celery-beat = "^2.8.1"
redis = "^6.2.0"
psycopg2-binary = "^2.9.10"
drf-spectacular = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
